seperated objects from functions
